Move code from rest-common to submit-api
