--- rest-common/cardano-rest-common.cabal
@@ -1,50 +1,0 @@
-cabal-version:          >= 1.10
-
--- http://haskell.org/cabal/users-guide/
-
-name:                   cardano-rest-common
-version:                3.1.2
-synopsis:               A Block Explorer for the Cardano network
-description:
-homepage:               https://github.com/input-output-hk/cardano-explorer
-bug-reports:            https://github.com/input-output-hk/cardano-explorer/issues
-license:                Apache-2.0
-license-file:           LICENSE
-author:                 IOHK Engineering Team
-maintainer:             operations@iohk.io
-copyright:              (c) 2019-2021 IOHK
-category:               Language
-build-type:             Simple
-extra-source-files:     CHANGELOG.md
-
-library
-  default-language:     Haskell2010
-  hs-source-dirs:       src
-  ghc-options:          -Wall
-                        -Werror
-                        -Wcompat
-                        -fwarn-redundant-constraints
-                        -fwarn-incomplete-patterns
-                        -fwarn-unused-imports
-                        -Wincomplete-record-updates
-                        -Wincomplete-uni-patterns
-                        -Wno-all-missed-specialisations
-                        -Wno-implicit-prelude
-                        -Wno-missing-import-lists
-                        -Wno-safe
-                        -Wno-unsafe
-
-  exposed-modules:      Cardano.Rest.Parsers
-                      , Cardano.Rest.Types
-                      , Cardano.Rest.Web
-
-  -- other-modules:
-
-  build-depends:        base                            >= 4.12         && < 4.13
-                      , optparse-applicative
-                      , monad-logger
-                      , text
-                      , streaming-commons
-                      , network
-                      , servant-server
-                      , warp
